+++ unit_test.py
@@ -1,0 +1,12 @@
+
+def get_units():
+    ret = []
+    # n, s = 25, 10
+    ret.append(([{4}, {2, 3}, {1, 4, 5, 8}, {1, 8}, {0, 2, 5}, {2, 4, 7, 8, 9, 10, 12, 14}, {11, 13, 14}, {5, 10}, {2, 3, 5, 12, 15, 17}, {5, 11, 13, 15, 18}, {5, 7, 17}, {6, 9, 13, 14, 15, 19}, {5, 8, 13, 14, 15, 19, 20}, {6, 9, 11, 12, 17, 19, 20, 22, 23}, {5, 6, 11, 12, 16, 18, 19, 22}, {8, 9, 11, 12, 18, 23, 24}, {14, 20, 21}, {8, 10, 13, 19, 21, 22}, {9, 14, 15, 20, 21}, {11, 12, 13, 14, 17, 21, 24}, {12, 13, 16, 18}, {16, 17, 18, 19, 22, 24}, {13, 14, 17, 21}, {13, 15}, {15, 19, 21}], 18))
+    ret.append(([{3}, {5, 6, 8, 11}, {8, 12}, {0, 4, 7, 8, 12}, {8, 3}, {1, 6, 8, 13}, {1, 5, 7, 11}, {3, 6, 9, 10}, {1, 2, 3, 4, 5, 14, 15, 17}, {7, 16, 17}, {7, 12, 15, 18}, {1, 6, 13, 14, 15}, {2, 3, 10, 17}, {5, 11, 17, 21, 22}, {8, 11, 15, 16, 22, 24}, {8, 10, 11, 14, 19, 20, 23}, {9, 14, 23}, {8, 9, 12, 13}, {10, 23, 24}, {15, 21, 22}, {15, 23, 24}, {13, 19}, {13, 14, 19, 23}, {15, 16, 18, 20, 22, 24}, {14, 18, 20, 23}], 14))
+    ret.append(([set(), {5, 9, 11}, {4, 5}, {11}, {2, 10, 11}, {1, 2, 7, 10, 11}, {8, 10, 16}, {5, 13, 15, 16}, {6, 10, 11, 13}, {1, 12, 14, 19}, {4, 5, 6, 8, 11, 13, 18}, {1, 3, 4, 5, 8, 10, 14, 16, 18, 20, 21}, {9, 14, 15, 16, 22}, {7, 8, 10, 14, 17, 20, 23}, {9, 11, 12, 13, 22, 24}, {7, 12, 24}, {6, 7, 11, 12, 18, 20, 21, 23}, {19, 20, 13, 22}, {10, 11, 16, 20}, {9, 17, 21}, {11, 13, 16, 17, 18, 23}, {11, 16, 19}, {12, 14, 17}, {13, 16, 20}, {14, 15}], 16))
+    ret.append(([{2, 5, 9}, {2, 4, 8}, {0, 1, 3, 8}, {10, 2}, {1, 6, 7}, {0, 7, 8, 11, 12, 13, 14}, {4}, {4, 5, 10, 12, 13, 17}, {1, 2, 5}, {0, 11, 12, 13, 14, 15, 19}, {3, 7, 15, 20}, {5, 9}, {5, 7, 9, 20}, {5, 7, 9, 16, 23}, {5, 9, 17, 21, 24}, {9, 10, 19, 20, 21, 24}, {13, 18, 19}, {7, 14}, {16}, {9, 15, 16}, {10, 12, 15, 21, 22}, {14, 15, 20}, {20, 23, 24}, {13, 22, 24}, {14, 15, 22, 23}], 12))
+    ret.append(([{1, 4, 6, 9}, {0, 5, 8, 11}, {10, 4}, {5, 7, 9}, {0, 2, 9, 11}, {1, 3, 6, 9, 13, 15}, {0, 5, 7, 8, 9, 13, 14, 15}, {3, 6, 9}, {1, 6, 9, 16, 17}, {0, 3, 4, 5, 6, 7, 8, 10, 11, 12, 13}, {2, 9, 11, 15, 19, 20}, {1, 4, 9, 10, 12}, {9, 11, 14, 16, 21}, {5, 6, 9}, {6, 12, 17, 22}, {5, 6, 10, 19}, {8, 12, 19, 20, 21}, {8, 14}, set(), {10, 15, 16, 24}, {10, 16, 21, 23}, {12, 16, 20, 22}, {14, 21}, {20}, {19}], 16))
+    # 40, 10
+
+    return ret
